Extracted tests requiring firefox into BrowserTest
